pyproject.toml
[tool.poetry]
name = "crudslong"
version = "0.1.0"
description = ""
authors = ["4utomator <mohammad@stilyoapps.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.4"}
asyncpg = "^0.30.0"
pydantic-settings = "^2.6.1"
sqlmodel = "^0.0.22"
pytest = "^8.3.3"
requests = "^2.32.3"
alembic = "^1.14.0"
passlib = "^1.7.4"
bcrypt = "^4.2.0"
pyjwt = "^2.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
